+++ libs/riscv/src/sbi/rfence.rs
@@ -32,7 +32,7 @@ pub fn sfence_vma_asid(
     hart_mask_base: usize,
     start_addr: usize,
     size: usize,
-    asid: usize,
+    asid: u16,
 ) -> super::Result<()> {
     sbi_call!(ext: EID_RFNC, func: 2, "a0": hart_mask, "a1": hart_mask_base, "a2": start_addr, "a3": size, "a4": asid)?;
     Ok(())
